Improves remove button, adds removeFeed function
Fixes and improves the way the remove button and the way the refresh button works. Adds the removeFeed function, creates a new window and populates it with all the feeds found on file. Clicking a button removes the feed from the urls list and skips it when writing to file. This removes the entry in sites.txt.